feat(config) live-Channels entries list + creation view default config on build
